refactoring code
changed the button and states position to modal testing page component.
