+++ client/src/pages/ModalTestingPage.js
@@ -1,7 +1,11 @@
-import React from "react";
+import React, { useState } from "react";
+import Button from "react-bootstrap/Button";
 import ModalVideo from "../components/ModalVideo";
 
 function ModalTestingPage(){
+    const [show, setShow] = useState(false);
+    const handleClose = () => setShow(false);
+    const handleShow = () => setShow(true);
     const data =
         {
             title: "Rumba Guaguanco",
@@ -11,7 +15,10 @@ function ModalTestingPage(){
 
     return <div>
         <h2>Testing Modal</h2>
-        <ModalVideo title={data.title} url={data.url} location={data.location} />
+        <Button variant="primary" onClick={handleShow}>
+            Explore Rhythm
+        </Button>
+        <ModalVideo show={show} handleClose={handleClose} title={data.title} url={data.url} location={data.location} />
 
         </div>;
